\docType{package}
\name{FLData-package}
\alias{FLData}
\alias{FLData-package}
\title{Tests pseudoprimality by Fermat's little theorem}
\description{
  A simple and robust statistical catch at age model
}
\details{
  \tabular{ll}{ Package: \tab FLData \cr Type: \tab Package
  \cr Version: \tab 0.1 \cr Date: \tab 2012-08-22 \cr
  License: \tab GPL (>= 2) \cr LazyLoad: \tab yes \cr }
}
\author{
  Ernesto Jardim \email{ernesto.jardim@jrc.ec.europa.eu}

  Iago Mosqueira \email{iago.mosqueira@jrc.ec.europa.eu}

  Chato Osio \email{chato.osio@jrc.ec.europa.eu}

  Colin Millar \email{colin.millar@jrc.ec.europa.eu}
}
\references{
  none
}
\seealso{
  \code{\link{paste}}
}
\keyword{package}

